first training script and adaptations
