+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/maelle/miniconda3/envs/neurodash/bin/python"
+}
